+++ derive/tests/union.rs
@@ -1,5 +1,6 @@
 mod schema_utils;
 mod union {
     mod union_tests;
+    mod with_generic_tests;
     mod with_interface;
 }